Don't cast `startTime` to `float` when initializing `passedTime`

+++ src/ru/nsu/ccfit/zuev/osu/game/GameplaySlider.java
@@ -167,7 +167,7 @@ public class GameplaySlider extends GameObject {
         position.set(stackedPosition.x, stackedPosition.y);
 
         endsCombo = beatmapSlider.isLastInCombo();
-        passedTime = secPassed - (float) beatmapSlider.startTime / 1000;
+        passedTime = secPassed - beatmapSlider.startTime / 1000;
         spanDuration = beatmapSlider.getSpanDuration() / 1000;
         path = sliderPath;
